+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @lottiefiles/dotlottie-svelte
 
+## 0.3.4
+
+### Patch Changes
+
+- Updated dependencies [5b942aa]
+  - @lottiefiles/dotlottie-web@0.25.0
+
+## 0.3.3
+
+### Patch Changes
+
+- Updated dependencies [663fab2]
+- Updated dependencies [b72a4d7]
+  - @lottiefiles/dotlottie-web@0.24.0
+
 ## 0.3.2
 
 ### Patch Changes